requirements.txt
uvloop
PyYAML
psutil
cx_Oracle
python-ldap